Fixed IntelliJ warnings

+++ src/main/java/timelessodyssey/view/screens/CreditsViewer.java
@@ -13,7 +13,6 @@ import java.io.IOException;
 import static timelessodyssey.view.text.GameTextViewer.*;
 
 public class CreditsViewer extends ScreenViewer<Credits> {
-
     private final TextViewer textViewer;
     private final LogoViewer logoViewer;
 
@@ -51,7 +50,7 @@ public class CreditsViewer extends ScreenViewer<Credits> {
             String message = getModel().getMessages()[idx];
             int messageLength = message.length() * getCharWidth() + (message.length() - 1) * getSpacing();
             textViewer.draw(message,
-                    (gui.getWidth() / 2) - (messageLength / 2),
+                    ((double) gui.getWidth() / 2) - ((double) messageLength / 2),
                     yAlignment + spacing * idx,
                     messageColor, gui);
         }
